Working across multiple pagesw on an Uno
Updated the thermo demo, EPD and EPD_GFX to allow a mode that will work
on Arduino Uno.

The page is divided vertically into a subset -- say 8 -- and each of
those is created/updated separately in a loop that re-executes each
graphic display against each page (and only applies pixel changes inside
the page).